refactor:  move some hooks
