+++ packages/web/src/components/table-handle/hooks/use-empty-image.ts
