+++ packages/web/src/components/table-handle/table-handle-row-trigger/setup.ts
@@ -7,11 +7,11 @@ import {
 import { useMenuTrigger } from '@aria-ui/menu/elements'
 import { selectTableRow } from '@prosekit/extensions/table'
 
-import { useEmptyImage } from '../../../hooks/use-empty-image'
 import {
   tableHandleDndContext,
   tableHandleRootContext,
 } from '../context'
+import { useEmptyImage } from '../table-handle-column-trigger/use-empty-image'
 
 import type {
   TableHandleRowTriggerEvents,
